pyproject.toml
[tool.poetry]
name = "django_hierarchical_files"
version = "0.1.0"
description = ""
authors = ["Cedric Mulvihill <cedric.mulvihill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
